finished all with views
