+++ src/daos/product_category_dao.py
@@ -9,7 +9,10 @@ set the queries to save in product_category table on database.
 
 class ProductCategoryDao(Dao):
 
-    def create_table_product_category(self):
+    def __init__(self):
+        self.create_table()
+
+    def create_table(self):
         self.execute_query("""
         CREATE TABLE IF NOT EXISTS product_category (
             product_id INTEGER,
